requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
langchain==0.0.350
langchain-openai==0.0.2
langsmith==0.0.69
python-dotenv==1.0.0
pydantic==2.5.0
